Make LXD buildable on 32bit arches again

+++ shared/util.go
@@ -504,7 +504,7 @@ func RunningInUserNS() bool {
 	}
 
 	line := string(l)
-	var a, b, c int
+	var a, b, c int64
 	fmt.Sscanf(line, "%d %d %d", &a, &b, &c)
 	if a == 0 && b == 0 && c == 4294967295 {
 		return false
